+++ R/chat.R
@@ -665,9 +665,10 @@ Chat <- R6::R6Class(
       } else {
         turn <- value_turn(
           private$provider,
-          response,
+          resp_body_json(response),
           has_type = !is.null(type)
         )
+        turn@duration <- resp_timing(response)[["total"]] %||% NA_real_
         turn <- match_tools(turn, private$tools)
 
         text <- turn@text
@@ -745,7 +746,12 @@ Chat <- R6::R6Class(
       } else {
         result <- await(response)
 
-        turn <- value_turn(private$provider, result, has_type = !is.null(type))
+        turn <- value_turn(
+          private$provider,
+          resp_body_json(result),
+          has_type = !is.null(type)
+        )
+        turn@duration <- resp_timing(result)[["total"]] %||% NA_real_
         text <- turn@text
         if (!is.null(text)) {
           emit(text)